requirements.txt
pypdf>=6.4.0
pandas>=1.5.0
numpy>=1.26.0
scipy>=1.11.0
sentence-transformers>=3.0.1
spacy==3.7.4
faiss-cpu>=1.8.0
torch>=2.0.0
openpyxl>=3.1.5
google-generativeai>=0.8.3
